% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{mcom_cols}
\alias{mcom_cols}
\title{Function to extract Mediacom colors as hex codes}
\usage{
mcom_cols(...)
}
\arguments{
\item{...}{Character names of mcom_colors}
}
\description{
Function to extract Mediacom colors as hex codes
}
\examples{
ggplot2::ggplot(data=mtcars, ggplot2::aes(x=disp, y=mpg)) +
  ggplot2::geom_point(colour = mcom_cols("signature pink"))

}
