% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{mcom_palette}
\alias{mcom_palette}
\title{Function to provide a standard MediaCom colour palette}
\usage{
mcom_palette()
}
\value{
\code{vector} of hex codes
}
\description{
Function to provide a standard MediaCom colour palette
}
\examples{
mcom_palette()

ggplot2::ggplot(data=mtcars[1:5,], ggplot2::aes(x=rownames(mtcars)[1:5], y=mpg, fill=rownames(mtcars)[1:5])) +
  ggplot2::geom_col() +
  ggplot2::scale_fill_manual(values = mcom_palette()) +
  mcom_ggplot(title = "My Chart", subtitle = "My Subtitle", xtitle = "X Axis", ytitle = "Y Axis", source = "My Source")
}
