% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{mcom_palette_primary}
\alias{mcom_palette_primary}
\title{Function to provide a standard MediaCom colour palette - primary colours}
\usage{
mcom_palette_primary()
}
\value{
\code{vector} of hex codes
}
\description{
Function to provide a standard MediaCom colour palette - primary colours
}
\examples{
mcom_palette_primary()

ggplot2::ggplot(data=mtcars[1:5,], ggplot2::aes(x=rownames(mtcars)[1:5], y=mpg, fill=rownames(mtcars)[1:5])) +
  ggplot2::geom_col() +
  ggplot2::scale_fill_manual(values = mcom_palette_primary())
}
