% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{mcom_palette_secondary}
\alias{mcom_palette_secondary}
\title{Function to provide a standard MediaCom colour palette - secondary colours}
\usage{
mcom_palette_secondary()
}
\value{
\code{vector} of hex codes
}
\description{
Function to provide a standard MediaCom colour palette - secondary colours
}
\examples{
mcom_palette_secondary()

ggplot2::ggplot(data=mtcars[1:5,], ggplot2::aes(x=rownames(mtcars)[1:5], y=mpg, fill=rownames(mtcars)[1:5])) +
  ggplot2::geom_col() +
  ggplot2::scale_fill_manual(values = mcom_palette_secondary())
}
