% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{pal_len}
\alias{pal_len}
\title{Return length of given palette}
\usage{
pal_len(palette)
}
\arguments{
\item{palette}{Character name of palette in mcom_palettes}
}
\value{
\code{numeric} number of colours in palette
}
\description{
Return length of given palette
}
\examples{
pal_len("main")

}
