% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_manual.R
\name{scale_color_mcom}
\alias{scale_color_mcom}
\title{Color scale constructor for Mediacom colors}
\usage{
scale_color_mcom(
  palette = "main",
  n = pal_len(palette),
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for Mediacom colors
}
\examples{
ggplot2::ggplot(data=mtcars, ggplot2::aes(x=disp, y=mpg,
                                          colour = as.factor(carb))) +
  ggplot2::geom_point() +
  # n needs to be the same as number of groups colouring by
  scale_color_mcom(palette = "secondary", n = 6) +
  theme_mcom()

# with a continuos scale
ggplot2::ggplot(data=mtcars, ggplot2::aes(x=disp, y=mpg, colour = wt)) +
  ggplot2::geom_point() +
  scale_color_mcom(discrete = F)

}
