% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_manual.R
\name{scale_fill_mcom}
\alias{scale_fill_mcom}
\title{Fill scale constructor for Mediacom colors}
\usage{
scale_fill_mcom(
  palette = "main",
  n = pal_len(palette),
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for Mediacom colors
}
\examples{
ggplot2::ggplot(data=mtcars, ggplot2::aes(x= as.factor(cyl), y=mean(mpg),
                                          fill = as.factor(gear))) +
  ggplot2::geom_col() +
  scale_fill_mcom(palette = "secondary", n = 3) +
  theme_mcom()
}
