% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veiw_palette.R
\name{view_palette}
\alias{view_palette}
\title{Allows you to view all mediacom palettes and names used in mcom_pal or mcom_pal_for_scale}
\usage{
view_palette(palette = "all", n = pal_len(palette))
}
\arguments{
\item{palette}{Character name of palette to view. If "all" then all palettes are shown}

\item{n}{can spescify number of colours you want to use. if n > palette length then more
colours are interpolated}
}
\value{
\code{list} of ggplot2 formatting options
}
\description{
Allows you to view all mediacom palettes and names used in mcom_pal or mcom_pal_for_scale
}
\examples{

view_palette()
view_palette("main")
view_palette("main", 2)

}
