% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mcom.R
\name{theme_mcom}
\alias{theme_mcom}
\title{Re-skins a ggplot2 chart with MediaCom titles and axes}
\usage{
theme_mcom(
  title = element_blank(),
  subtitle = element_blank(),
  xtitle = element_blank(),
  ytitle = element_blank(),
  source = "MediaCom"
)
}
\arguments{
\item{title}{Chart title}

\item{subtitle}{Chart subtitle}

\item{xtitle}{X axis label}

\item{ytitle}{Y axis label}

\item{source}{Data source to display in bottom right}
}
\value{
\code{list} of ggplot2 formatting options
}
\description{
Re-skins a ggplot2 chart with MediaCom titles and axes
}
\examples{
ggplot2::ggplot(data=mtcars, ggplot2::aes(x=disp, y=mpg)) +
 ggplot2::geom_line(colour = mcom_signature_Pink()) +
 theme_mcom(title = "My Chart", subtitle = "My Subtitle", xtitle = "X Axis", ytitle = "Y Axis", source = "My Source")
}
